+++ packages/interbit-cli/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.4.20"></a>
+## [0.4.20](https://github.com/interbit/interbit/compare/v0.4.19...v0.4.20) (2018-05-11)
+
+
+
+
+**Note:** Version bump only for package interbit-cli
+
+<a name="0.4.19"></a>
+## [0.4.19](https://github.com/interbit/interbit/compare/v0.4.18...v0.4.19) (2018-05-11)
+
+
+
+
+**Note:** Version bump only for package interbit-cli
+
 <a name="0.4.18"></a>
 ## [0.4.18](https://github.com/interbit/interbit/compare/v0.4.17...v0.4.18) (2018-05-10)